requirements.txt
amqp==5.0.9
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.1
certifi==2022.12.7
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
boto3==1.24.36
Django~=4.1.5
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-js-asset==1.2.2
django-storages==1.11.1
djangorestframework==3.13.1
filesplit==4.0.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.2
mccabe==0.6.1
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pyflakes==2.4.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.3
requests==2.26.0
setuptools==63.2.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Unidecode==1.3.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wheel==0.38.4
XlsxWriter==3.0.2
xlrd==2.0.1
lxml
html5lib
bs4
channels~=4.0.0
channels-redis~=4.0.0
pandas~=1.5.3

numpy~=1.23.1
cffi~=1.15.1
asgiref~=3.6.0